Aleksandrs K.: rename package from level 2_3 to level 2_3_4
