+++ src/main/java/student_aleksandrs_korsaks/ak_lesson_10/level_2_3_4/AndSearchCriteria.java
@@ -1,4 +1,4 @@
-package student_aleksandrs_korsaks.ak_lesson_10.level_2_3;
+package student_aleksandrs_korsaks.ak_lesson_10.level_2_3_4;
 
 class AndSearchCriteria implements SearchCriteria {
