+++ src/main/java/student_aleksandrs_korsaks/ak_lesson_10/level_2_3_4/BookDatabaseImpl.java
@@ -1,8 +1,9 @@
-package student_aleksandrs_korsaks.ak_lesson_10.level_2_3;
+package student_aleksandrs_korsaks.ak_lesson_10.level_2_3_4;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
 
 class BookDatabaseImpl implements BookDatabase {
 
@@ -73,5 +74,10 @@ class BookDatabaseImpl implements BookDatabase {
         return bookList.stream()
                 .filter(searchCriteria::match)
                 .toList();
+    }
+
+    @Override
+    public Set<String> findUniqueAuthors() {
+        return null;
     }
 }
